# tests/fixtures/assistant_openapi_baseline.json
{
  "components": {
    "schemas": {
...
      "CapabilityCatalogRequest": {
        "properties": {
          "capability_revision": {
            "minimum": 1.0,
            "title": "Capability Revision",
            "type": "integer"
          },
          "model_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "session_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "tenant_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Tenant Id",
            "type": "string"
          },
          "user_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "user_id",
          "session_id",
          "capability_revision",
          "model_id"
        ],
        "title": "CapabilityCatalogRequest",
        "type": "object"
      },
      "CapabilityInvokeRequest": {
        "properties": {
          "arguments": {
            "additionalProperties": true,
            "title": "Arguments",
            "type": "object"
          },
          "bound_dataset_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Bound Dataset Ids",
            "type": "array"
          },
          "capability_revision": {
            "minimum": 1.0,
            "title": "Capability Revision",
            "type": "integer"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          },
          "session_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "snapshot_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Snapshot Id",
            "type": "string"
          },
          "tenant_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Tenant Id",
            "type": "string"
          },
          "tool": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:/-]+$",
            "title": "Tool",
            "type": "string"
          },
          "user_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "user_id",
          "session_id",
          "capability_revision",
          "snapshot_id",
          "run_id",
          "tool"
        ],
        "title": "CapabilityInvokeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI Assistant Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/capabilities/catalog": {
      "post": {
        "operationId": "capability_catalog_internal_v1_capabilities_catalog_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityCatalogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capability Catalog",
        "tags": [
          "Internal Capabilities"
        ]
      }
    },
    "/internal/v1/capabilities/invoke": {
      "post": {
        "operationId": "capability_invoke_internal_v1_capabilities_invoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityInvokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capability Invoke",
        "tags": [
          "Internal Capabilities"
        ]
      }
    }
  }
}